refactor(contracts): organize command schemas by domain
