+++ packages/contracts/src/commands/source.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+import { nonEmptyStringSchema } from "../shared.js";
+
+export const CommandSourceSchema = z
+  .object({
+    kind: z.enum(["branch", "pr", "manual"]),
+    value: nonEmptyStringSchema,
+  })
+  .strict();
