+++ packages/contracts/src/index.ts
@@ -1,4 +1,12 @@
 export * from "./agentStatus.js";
+export * from "./commandLifecycle.js";
+export * from "./commands/observer.js";
+export * from "./commands/project.js";
+export * from "./commands/session.js";
+export * from "./commands/sessionGroup.js";
+export * from "./commands/source.js";
+export * from "./commands/terminal.js";
+export * from "./commands/worktree.js";
 export * from "./commands.js";
 export * from "./diagnostics.js";
 export * from "./errors.js";
